backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-cors==5.0.0
flask-restful==0.3.9
flask-sqlalchemy==2.5.1
openai==0.27.8
Flask-WTF==1.2.1
WTForms==3.1.2
Flask-Uploads==0.2.1
bcrypt==4.2.0
flask-bcrypt==1.0.1
Flask-Mail==0.9.1
blinker==1.8.2
python_dotenv==1.0.1
pandas==2.2.3